implement partial role-based access control

+++ lottery/views.py
@@ -3,6 +3,7 @@ import copy
 import logging
 
 from flask import Blueprint, render_template, request, flash
+from flask_login import login_required, current_user
 
 from app import db
 from models import Draw, User
@@ -11,18 +12,16 @@ from models import Draw, User
 lottery_blueprint = Blueprint('lottery', __name__, template_folder='templates')
 
 
-user = User.query.first()
-draw_key = user.draw_key
-
-
 # VIEWS
 # view lottery page
 @lottery_blueprint.route('/lottery')
+@login_required
 def lottery():
     return render_template('lottery.html')
 
 
 @lottery_blueprint.route('/add_draw', methods=['POST'])
+@login_required
 def add_draw():
     submitted_draw = ''
     for i in range(6):
@@ -30,7 +29,7 @@ def add_draw():
     submitted_draw.strip()
 
     # create a new draw with the form data.
-    new_draw = Draw(user_id=1, draw=submitted_draw, win=False, round=0, draw_key=draw_key)  # TODO: update user_id [user_id=1 is a placeholder]
+    new_draw = Draw(user_id=current_user.id, draw=submitted_draw, win=False, round=0, draw_key=current_user.draw_key)
 
     # add the new draw to the database
     db.session.add(new_draw)
@@ -43,9 +42,10 @@ def add_draw():
 
 # view all draws that have not been played
 @lottery_blueprint.route('/view_draws', methods=['POST'])
+@login_required
 def view_draws():
     # get all draws that have not been played [played=0]
-    playable_draws = Draw.query.filter_by(played=False).all()  # TODO: filter playable draws for current user
+    playable_draws = Draw.query.filter_by(user_id=current_user.id, played=False).all()
 
     # if playable draws exist
     if len(playable_draws) != 0:
@@ -54,7 +54,7 @@ def view_draws():
         decrypted_draws = []
 
         for d in draw_copies:
-            d.view_draw(draw_key)
+            d.view_draw(current_user.draw_key)
             decrypted_draws.append(d)
 
         return render_template('lottery.html', playable_draws=decrypted_draws)
@@ -65,9 +65,10 @@ def view_draws():
 
 # view lottery results
 @lottery_blueprint.route('/check_draws', methods=['POST'])
+@login_required
 def check_draws():
     # get played draws
-    played_draws = Draw.query.filter_by(played=True).all()  # TODO: filter played draws for current user
+    played_draws = Draw.query.filter_by(user_id=current_user.id, played=True).all()
 
     # if played draws exist
     if len(played_draws) != 0:
@@ -75,7 +76,7 @@ def check_draws():
         decrypted_draws = []
 
         for d in draw_copies:
-            d.view_draw(draw_key)
+            d.view_draw(current_user.draw_key)
             decrypted_draws.append(d)
 
         return render_template('lottery.html', results=decrypted_draws, played=True)
@@ -88,8 +89,9 @@ def check_draws():
 
 # delete all played draws
 @lottery_blueprint.route('/play_again', methods=['POST'])
+@login_required
 def play_again():
-    delete_played = Draw.__table__.delete().where(Draw.played)  # TODO: delete played draws for current user only
+    delete_played = Draw.__table__.delete().where(Draw.user_id == current_user.id, Draw.played)
     db.session.execute(delete_played)
     db.session.commit()
